pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "review-buddy"
version = "2.0.0"
description = "Intelligent academic paper search, filtering, and download tool"
readme = "README.md"
requires-python = ">=3.7"
license = {text = "MIT"}
authors = [
    {name = "Review Buddy Contributors"}
]
keywords = ["academic", "papers", "research", "bibliography", "pdf", "download", "filter"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering",
]

dependencies = [
    "requests>=2.31.0",
    "lxml>=4.9.0",
    "python-dotenv>=1.0.0",
    "beautifulsoup4>=4.12.0",
    "tqdm>=4.66.0",
    "bibtexparser>=1.4.0",
    "rispy>=0.7.0",
    "langdetect>=1.0.9",
    "pandas>=1.3.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "streamlit>=1.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
scholar = [
    "scholarly>=1.7.0",
]
scihub = [
    "scihub",
]
all = [
    "review-buddy[dev,scholar,scihub]",
]

[project.urls]
Homepage = "https://github.com/leonardozaggia/review_buddy"
Documentation = "https://github.com/leonardozaggia/review_buddy/blob/main/README.md"
Repository = "https://github.com/leonardozaggia/review_buddy"
Issues = "https://github.com/leonardozaggia/review_buddy/issues"

[project.scripts]
reviewbuddy = "cli:main"

[tool.setuptools]
packages = ["core", "src", "src.searchers"]

[tool.black]
line-length = 100
target-version = ['py37', 'py38', 'py39', 'py310', 'py311']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=core --cov=src --cov-report=html --cov-report=term"

[tool.mypy]
python_version = "3.7"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
